// frontend/src/api/models/meeting-note-output-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface MeetingNoteOutputDTO
 */
export interface MeetingNoteOutputDTO {
    /**
     * 
     * @type {string}
     * @memberof MeetingNoteOutputDTO
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof MeetingNoteOutputDTO
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MeetingNoteOutputDTO
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MeetingNoteOutputDTO
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof MeetingNoteOutputDTO
     */
    'content'?: string;
}