// frontend/src/api/models/update-meeting-note-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateMeetingNoteDTO
 */
export interface UpdateMeetingNoteDTO {
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingNoteDTO
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingNoteDTO
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMeetingNoteDTO
     */
    'content'?: string;
}